move create_new_cell into utils
